+++ my-wellness-app/db.json
@@ -5,7 +5,7 @@
         "exercise_name": "Overhead Tricep Stretch",
         "muscle_group": "Triceps",
         "category": "arms",
-        "description": "Raise your arm and bend it so that your hand reaches toward the opposite side. Use your other hand and pull the elbow toward your head until you feel a stretch",
+        "description": ["Raise your arm and bend it so that your hand reaches toward the opposite side.", "Use your other hand and pull the elbow toward your head until you feel a stretch"],
         "duration": "Hold the stretch for 15-30 seconds on each arm",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Triceps_Stretch.gif",
         "Video":"test"
@@ -15,7 +15,7 @@
         "exercise_name": "Overhead Reach",
         "muscle_group": "Lats",
         "category": "back",
-        "description": "Extend your arm overhead, and reach to the opposite side",
+        "description": ["Extend your arm overhead, and reach to the opposite side"],
         "duration": "Hold the stretch for 15-30 seconds on each side",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Overhead_Reach.gif",
         "Video":"test"
@@ -25,7 +25,7 @@
         "exercise_name": "Overhead Stretch",
         "muscle_group": "Upper back, rotator cuff, etc",
         "category": "back",
-        "description":"Clasp hands together above the head with palms facing outward. Push your arms up, stretching upward.", 
+        "description":["Clasp hands together above the head with palms facing outward.", "Push your arms up, stretching upward."], 
         "duration": "Hold the stretch for 15-30 seconds on each arm",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Upper_Body_Stretch.gif",
         "Video":"test"
@@ -35,7 +35,7 @@
         "exercise_name": "Shoulder / Chest Stretch",
         "muscle_group": "Shoulders",
         "category": "shoulders",
-        "description": "Clasp hands behind your back, push your chest out and tilt your head upwards",
+        "description": ["Clasp hands behind your back", "Push your chest out and tilt your head upwards"],
         "duration": "Hold the stretch for 15-30 seconds",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Shoulder_Pec_Stretch.gif",
         "Video":"test"
@@ -45,7 +45,7 @@
         "exercise_name": "Forward Bend",
         "muscle_group": "Upper Back",
         "category": "back",
-        "description": "Clasp your hands in front of you, round your back forward and lower your head to be in line with your arms",
+        "description": ["Clasp your hands in front of you", "Round your back forward and lower your head to be in line with your arms"],
         "duration": "Hold the stretch for 15-30 seconds",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Forward_Stretch.gif",
         "Video":"test"
@@ -55,7 +55,7 @@
         "exercise_name": "Torso Rotation",
         "muscle_group": "abs",
         "category": "lower back",
-        "description": "Twist your upper body in the direction of the arm that's resting on the back of your chair",
+        "description": ["Twist your upper body in the direction of the arm that's resting on the back of your chair"],
         "duration": "Hold the stretch for 15-30 seconds on each side",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Torso_Stretch.gif",
         "Video":"test"
@@ -65,7 +65,7 @@
         "exercise_name": "Toe Touches",
         "muscle_group": "hamstrings",
         "category": "legs",
-        "description": "Raise your arm and bend it so that your hand reaches toward the opposite side. Use your other hand and pull the elbow toward your head until you feel a stretch",
+        "description": ["Raise your arm and bend it so that your hand reaches toward the opposite side.", "Use your other hand and pull the elbow toward your head until you feel a stretch"],
         "duration": "Hold the stretch for 15-30 seconds on each arm",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Hamstring_Stretch.gif",
         "Video":"test"
@@ -75,7 +75,7 @@
         "exercise_name": "Upper Trap Stretch",
         "muscle_group": "Traps",
         "category": "neck",
-        "description": "Gently pull your head toward each shoulder until you feel a light stretch",
+        "description": ["Gently pull your head toward each shoulder until you feel a light stretch"],
         "duration": "Hold the stretch for 15-30 seconds on each side",
         "exercise_image":"https://images-prod.healthline.com/hlcmsresource/images/topic_centers/Fitness-Exercise/400x400_Stretches_to_Do_at_Work_Every_Day_Upper_Trap_Stretch.gif",
         "Video":"test"
